refactor(appkit): tidy stream cleanup helpers and de-dup terminal paths
- extract shared `_finalizeStream()` from the three identical generator
  terminal sequences (success, client-abort, error)
- move `clearGraceTimer`/`clearRemovalTimer` out of `types.ts` into a
  dedicated `stream/timers.ts` so `types.ts` stays types-only
- call `clearRemovalTimer()` in `_scheduleRemovalAfterTTL` instead of
  inlining `clearTimeout` (matches `_scheduleGraceAbort`)
- unexport internal `RingBuffer` and drop its knip suppression
